+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Added support for row validation using XSD schema using `rowValidationXSDPath` option when reading XML files with a row tag using `rowTag` option.
 - Improved SQL generation for `session.table().sample()` to generate a flat sql statement.
+- Added support for complex column expression as input for `functions.explode`.
 
 ### Snowpark Local Testing Updates